+++ cpp/include/wup/nodes/export.hpp
@@ -64,7 +64,7 @@ public:
     {
         std::ofstream file_out( _filename.c_str() );
         for (uint i=0; i<_bundle.numRows(); ++i) {
-            file_out << _bundle(i,0);
+            file_out << _bundle(i, uint(0));
             for (uint j=1; j<_bundle.numCols(); ++j)
                 file_out << "," << _bundle(i,j);
             file_out << std::endl;
